Whip project is now static! Log class has been changed a little(Some additions). Context classes have been added. Some warnings have been fixed.

+++ Whip/src/Whip/Events/Event Callback Functions/EventCallbackFunctions.cpp
@@ -48,7 +48,7 @@ void EventCallbackFunctions::KeyPressedReleasedEventsCallback(GLFWwindow* window
 
 		case GLFW_REPEAT:
 		{
-			KeyPressedEvent event(key, repeat_time++);
+			KeyPressedEvent event(key, (int)repeat_time++);
 			data.EventCallback(event);
 			break;
 		}
